GUI console refactor
Refactored console GUI elements for dedicated module
